+++ app/models/bitsy/blockchain_notification.rb
@@ -1,6 +1,8 @@
 module Bitsy
   class BlockchainNotification < ActiveRecord::Base
 
+    attr_reader :test
+
     validates(
       :value,
       :transaction_hash,
@@ -10,6 +12,10 @@ module Bitsy
     )
 
     validates(:secret, inclusion: {in: Bitsy.config.blockchain_secrets})
+
+    def test=(b)
+      @test = ["true", true].include?(b)
+    end
 
   end
 end