+++ spec/models/blockchain_notification_spec.rb
@@ -10,5 +10,15 @@ module Bitsy
       end
     end
 
+    describe "#test" do
+      it "is an accessor for testing purposes" do
+        notification = described_class.new(test: "true")
+        expect(notification.test).to be true
+
+        notification = described_class.new(test: "false")
+        expect(notification.test).to be false
+      end
+    end
+
   end
 end